Revert "Moved Discord bot to separate class"
This reverts commit 38f520388e32d3ab1f64d6a6be2ca3ddb624ac85.
